+++ src/main/java/com/example/lab5_team_log121/PerspectiveView.java
@@ -1,8 +1,11 @@
 package com.example.lab5_team_log121;
 
+import javafx.event.EventHandler;
 import javafx.geometry.Rectangle2D;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.MouseEvent;
+import javafx.scene.input.ScrollEvent;
 import javafx.scene.layout.StackPane;
 
 // Constructeur et méthodes de mise à jour pour la vue de perspective.
@@ -10,6 +13,9 @@ public class PerspectiveView extends StackPane implements Observer {
     private Perspective perspective;
     private ImageModel imageModel;
     private ImageView imageView;
+    
+    private double dragStartX, dragStartY;
+    private double initialOffsetX, initialOffsetY;
 
     // Constructeur : initialise la vue de perspective, configure l'ImageView et attache les observateurs.
     public PerspectiveView(ImageModel imageModel, Perspective perspective) {
@@ -23,6 +29,7 @@ public class PerspectiveView extends StackPane implements Observer {
         perspective.attach(this);
         imageModel.attach(this);
         update(null, "Init");
+        this.attachHandlers();
     }
 
     // Retourne l'ImageView utilisé pour afficher la perspective.
@@ -79,5 +86,52 @@ public class PerspectiveView extends StackPane implements Observer {
         Rectangle2D viewport = new Rectangle2D(offsetX, offsetY, viewportWidth, viewportHeight);
         imageView.setViewport(viewport);
 
+    }
+
+    private void attachHandlers() {
+
+        // Gestionnaire d'événement pour le zoom via la molette de la souris (appel direct sur le modèle)
+        this.getImageView().setOnScroll(new EventHandler<ScrollEvent>() {
+            @Override
+            public void handle(ScrollEvent event) {
+                double deltaY = event.getDeltaY();
+                double factor = (deltaY > 0) ? 1.1 : 0.9;
+                double centerX = event.getX() * perspective.getScale() + perspective.getOffsetX();
+                double centerY = event.getY() * perspective.getScale() + perspective.getOffsetY();
+                perspective.zoom(factor, centerX, centerY);
+                event.consume();
+            }
+        });
+
+        // Gestionnaire d'événement : démarre le pan en enregistrant la position initiale
+        this.getImageView().setOnMousePressed(new EventHandler<MouseEvent>() {
+            @Override
+            public void handle(MouseEvent event) {
+                dragStartX = event.getSceneX();
+                dragStartY = event.getSceneY();
+                initialOffsetX = perspective.getOffsetX();
+                initialOffsetY = perspective.getOffsetY();
+            }
+        });
+
+        // Gestionnaire d'événement : fournit un feedback visuel pendant le pan
+        this.getImageView().setOnMouseDragged(new EventHandler<MouseEvent>() {
+            @Override
+            public void handle(MouseEvent event) {
+                double dx = event.getSceneX() - dragStartX;
+                double dy = event.getSceneY() - dragStartY;
+                PerspectiveView.this.updatePan(initialOffsetX - dx, initialOffsetY - dy);
+            }
+        });
+
+        // Gestionnaire d'événement : finalise le pan en appliquant la translation sur le modèle
+        this.getImageView().setOnMouseReleased(new EventHandler<MouseEvent>() {
+            @Override
+            public void handle(MouseEvent event) {
+                double dx = event.getSceneX() - dragStartX;
+                double dy = event.getSceneY() - dragStartY;
+                perspective.move(-dx, -dy);
+            }
+        });
     }
 }
